requirements.txt
asgiref==3.5.0
Django==5.0.6
django-cors-headers==3.10.1
djangorestframework==3.13.1
python-decouple==3.6
pytz==2021.1
sqlparse==0.4.2
requests==2.32.3
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.134.0
python-dotenv==1.0.1